requirements.txt
python-telegram-bot==20.7
gspread==5.12.0
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
oauth2client==4.1.3
requests==2.31.0
SpeechRecognition==3.10.0
pydub==0.25.1
